web/requirements.txt
uwsgi
simplejson
peewee
PyJWT
redis
PyMySQL
